refactor(client): move client/client.ts to client/index.ts (default folder import)
