+++ src/Pages/ProductsPage/ProductsPage.tsx
@@ -15,7 +15,7 @@ import { ReactComponent as SvgFilter } from '@/assets/icons/filter.svg'
 import { ReactComponent as SvgSort } from '@/assets/icons/sort.svg'
 import { ReactComponent as SvgClose } from '@/assets/icons/close.svg'
 import { ReactComponent as SvgArrow } from '@/assets/icons/arrow-right.svg'
-import { apiRoot } from '@/client/client'
+import { apiRoot } from '@/client'
 import useCategories from '@/hooks/useCategories'
 import Checkbox from '@/Components/UIKit/Checkbox/Checkbox'
 import { useInView } from 'react-intersection-observer'
